requirements.txt
absl-py
addict
altair
attrs
blinker
cachetools
captcha
certifi
charset-normalizer
click
colorama
coloredlogs
contourpy
cycler
filelock
flatbuffers
fonttools
gitdb
GitPython
google-auth
google-auth-oauthlib
grpcio
humanfriendly
idna
IMcore
importlib-metadata
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
Markdown
markdown-it-py
MarkupSafe
matplotlib>=3.7.1
mdurl
mpmath
networkx
numpy>=1.24.0
oauthlib
onnxruntime
opencv-python>=4.8.0
packaging
pandas
Pillow
protobuf
psutil
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pycocotools
pydeck
Pygments
pyparsing
pyreadline3
PySide6
PySide6-Addons
PySide6-Essentials
python-dateutil
pytz
PyYAML>=6.0
QtFusion
referencing
requests
requests-oauthlib
rich
rsa
scipy
seaborn
shiboken6
six
smmap
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
thop
toolz
torch>=1.13.1
torchaudio>=2.0.0
torchvision>=0.14.1
tornado
tqdm
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
Werkzeug
zipp
timm
einops
efficientnet_pytorch
dill
pathlib>=1.0.1
scikit-learn>=1.3.0
ultralytics>=8.0.202
